+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-04-26 18:20:46
-// Commit: perf(perf): optimize memory consumption in perf list viewport
+// Updated at: 2026-04-27 09:30:51
+// Commit: refactor(perf): extract reusable atomic widgets from perf
 
 class perfComponentState {
-  final String commitRef = '722';
-  final String lastMessage = 'perf(perf): optimize memory consumption in perf list viewport';
-  final DateTime timestamp = DateTime.parse('2026-04-26T18:20:46+05:00');
+  final String commitRef = '736';
+  final String lastMessage = 'refactor(perf): extract reusable atomic widgets from perf';
+  final DateTime timestamp = DateTime.parse('2026-04-27T09:30:51+05:00');
 }